+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Updated
 
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.3.0-alpha.3`
+
 ## [0.1.0-beta.12] - 2023-01-04
 
 ### Added