+++ CHANGELOG.md
@@ -14,7 +14,8 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.3.0-alpha.4`
+- `kotlin -> 1.8.0`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.4.0-alpha.4`
 
 ## [0.1.0-beta.12] - 2023-01-04